split db config into separate file
