+++ app/__init__.py
@@ -4,12 +4,14 @@ from flask import Flask, render_template, request, redirect
 from dotenv import load_dotenv
 import json
 from peewee import *
-import datetime
 from playhouse.shortcuts import model_to_dict
+import libgravatar
 import werkzeug
 from werkzeug.security import generate_password_hash, check_password_hash
-import libgravatar
+
 from flask_login import *
+
+from app.db import User, TimelinePost
 
 load_dotenv()
 app = Flask(__name__)
@@ -19,6 +21,7 @@ login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = "signin"
 
+# generate secret key
 app.secret_key = secrets.token_urlsafe(16)
 
 # store portfolio directory, then get json data path
@@ -27,48 +30,6 @@ dataPath = os.path.join(portfolio_dir, "static/data.json")
 
 data = open(dataPath)
 data = json.load(data)
-
-# if env variable TESTING is set to true, instantiate a in-memory db for testing purposes only
-if os.getenv("TESTING") == "true":
-    print("Running in test mode")
-    mydb = SqliteDatabase("file:memory?mode=memory&cache=shared", uri=True)
-else:
-    # for production, connect to real db specified by .env variables
-    mydb = MySQLDatabase(
-        os.getenv("MYSQL_DATABASE"),
-        user=os.getenv("MYSQL_USER"),
-        password=os.getenv("MYSQL_PASSWORD"),
-        host=os.getenv("MYSQL_HOST"),
-        port=3306,
-    )
-
-print(mydb)
-
-# peewee model for the timeline posts
-class TimelinePost(Model):
-    name = CharField()
-    email = CharField()
-    content = TextField()
-    created_at = DateTimeField(default=datetime.datetime.now)
-    avatar = CharField(default="Testing")
-
-    class Meta:
-        database = mydb
-
-
-# peewee model for the users
-class User(UserMixin, Model):
-    name = CharField(unique=True)
-    password = CharField()
-
-
-    class Meta:
-        database = mydb
-
-
-# connect to the database, and create tables using the above models
-mydb.connect()
-mydb.create_tables([TimelinePost, User])
 
 
 ##### FRONTEND ROUTES #####
@@ -120,9 +81,83 @@ def get_signup():
     return render_template(
         "signup.html", title="Sam Thibault - Sign up", url=os.getenv("URL")
     )
+##### END OF FRONTEND ROUTES #####
 
 
-##### Authentication #####
+##### TIMELINE API ROUTES #####
+# add a document by specifying field values in the request body
+@app.route("/api/timeline_post", methods=["POST"])
+def post_time_line_post():
+
+    # start by checking if the http request structure is correct
+    req = request
+
+    if "email" not in req.form:
+        return "Invalid email, please try again", 400
+    elif "content" not in req.form:
+        return "Invalid content, please try again", 400
+
+    # if it is, we can assign some variables
+    name = current_user.name
+    print(name)
+    email = req.form["email"]
+    print(email)
+    content = req.form["content"]
+    print(content)
+
+    # use libgravatar to find profile image link for the submitted email, default to basic avatar
+    avatar = libgravatar.Gravatar(email).get_image(default="mm")
+
+    # if the request body is formatted properly, and frontend form validation fails, ensure the fields are formatted properly here
+    if content == "":
+        return "Invalid content, please try again", 400
+    elif "@" not in email or "." not in email:
+        return "Invalid email, please try again", 400
+    elif name == "":
+        return "Invalid name, please try again", 400
+    else:
+        timeline_post = TimelinePost.create(
+            name=name, email=email, content=content, avatar=avatar
+        )
+        return model_to_dict(timeline_post)
+
+
+# get all documents
+@app.route("/api/timeline_post", methods=["GET"])
+def get_time_line_post():
+    return {
+        "timeline_posts": [
+            model_to_dict(p)
+            for p in TimelinePost.select().order_by(TimelinePost.created_at.desc())
+        ]
+    }
+
+
+# delete a document by id
+@app.route("/api/timeline_post", methods=["DELETE"])
+def delete_time_line_post():
+
+    # get the author name associated with the post id
+    idToDelete = request.form["id"]
+    try:
+        qry = TimelinePost.get(idToDelete == TimelinePost.id)
+    except:
+        return "Invalid ID, please try again", 400
+
+    postAuthor = qry.name
+
+    # if the author matches the current user's name (case insensitive), try deleting, and return result message
+    if postAuthor.lower() == current_user.name.lower() or current_user.name == "admin":
+        qry = TimelinePost.delete().where(TimelinePost.id == idToDelete)
+        qry.execute()
+        return "deleted: " + idToDelete, 200
+    else:
+        return "You do not have permission to delete this post.", 400
+
+##### END OF TIMELINE API ROUTES #####
+
+
+##### AUTHENTICATION #####
 # get user based on id
 @login_manager.user_loader
 def load_user(user_id):
@@ -177,82 +212,4 @@ def signout():
     logout_user()
     return redirect("/")
 
-
-##### API ROUTES #####
-# add a document by specifying field values in the request body
-@app.route("/api/timeline_post", methods=["POST"])
-def post_time_line_post():
-
-    # start by checking if the http request structure is correct
-    req = request
-
-    if "email" not in req.form:
-        return "Invalid email, please try again", 400
-    elif "content" not in req.form:
-        return "Invalid content, please try again", 400
-
-    # if it is, we can assign some variables
-    name = current_user.name
-    print(name)
-    email = req.form["email"]
-    print(email)
-    content = req.form["content"]
-    print(content)
-
-    # use libgravatar to find profile image link for the submitted email, default to basic avatar
-    avatar = libgravatar.Gravatar(email).get_image(default="mm")
-
-    # if the request body is formatted properly, and frontend form validation fails, ensure the fields are formatted properly here
-    if content == "":
-        return "Invalid content, please try again", 400
-    elif "@" not in email or "." not in email:
-        return "Invalid email, please try again", 400
-    elif name == "":
-        return "Invalid name, please try again", 400
-    else:
-        timeline_post = TimelinePost.create(
-            name=name, email=email, content=content, avatar=avatar
-        )
-        return model_to_dict(timeline_post)
-
-
-# get all documents
-@app.route("/api/timeline_post", methods=["GET"])
-def get_time_line_post():
-    return {
-        "timeline_posts": [
-            model_to_dict(p)
-            for p in TimelinePost.select().order_by(TimelinePost.created_at.desc())
-        ]
-    }
-
-
-# delete a document by name
-@app.route("/api/timeline_post", methods=["DELETE"])
-def delete_time_line_post():
-
-    # get the author name associated with the post id
-    idToDelete = request.form["id"]
-    try:
-        qry = TimelinePost.get(idToDelete == TimelinePost.id)
-    except:
-        return "Invalid ID, please try again", 400
-
-    postAuthor = qry.name
-
-    # if the author matches the current user's name (case insensitive), try deleting, and return result message
-    if postAuthor.lower() == current_user.name.lower() or current_user.name == "admin":
-        qry = TimelinePost.delete().where(TimelinePost.id == idToDelete)
-        qry.execute()
-        return "deleted: " + idToDelete, 200
-    else:
-        return "You do not have permission to delete this post.", 400
-        
-
-
-# this mapping deletes all documents from the database (used for testing)
-@app.route("/api/timeline_post/purge", methods=["DELETE"])
-def delete_all():
-    qry = TimelinePost.delete()
-    qry.execute()
-    return "deleted all rows"
+##### END OF AUTHENTICATION #####
